+++ CHANGELOG.md
@@ -11,10 +11,13 @@ Make sure to update your `package.json` with following version constraints:
 }
 ```
 
-Note that 6.2.0 introduced several breaking changes to JS APIs:
+Other updates:
 
-- `QuerySnapshot.docChanges` is no more a field but a method
-  `QuerySnapshot.docChanges()`.
+- Added `Firestore.getAll()` method.
+
+Note that 6.2.0 of JS SDK introduced several breaking changes to JS APIs:
+
+- `QuerySnapshot.docChanges` is no more a field but a method `QuerySnapshot.docChanges()`.
 - For query snapshots returned from `DocumentQuery.get()` all `DocumentChange`s now return their
   type as `added` instead of `null`.
 - `GeoPoint` no longer exposes `toString` method.